Support new format of specifying env vars
Before this commit, the only way of setting up ENV variables was to use
env key in travis.yml, for example:

env:
  - FOO=bar
  - BAR=baz

All the elements are then expanded into matrix, so such env array will
result in 2 tests runs. With FOO=bar and then with BAR=baz ENV
variables. The problem with such approach is that in order to specify
ENV variables that are available globally, you need to add such vars for
each line, for example:

env:
  - FOO=bar SOME_TOKEN=abcdef
  - BAR=baz SOME_TOKEN=abcdef

New format helps in such situations by allowing to specify lines that go
into matrix and lines that will be available globally:

env:
  global:
    - SOME_TOKEN=abcdef
  matrix:
    - FOO=bar
    - BAR=baz