Ensure that duplicate, private repositories aren't synced.
This happens in odd circumstances: a user forks a repository of his/her
organization, and his API access lists the repository correctly. But it's
also included in the organization's (of which he's a member and who the
repository that was forked belongs to) list of repositories of this user.
The problem is that the organization's list shows lower permissions than the
user's, therefore it overwrites the original permissions and takes away admin
rights.

When it's intentional on the GitHub API's logic or not, this change discards
organization repositories that have already been seen via another resource,
e.g. the user's.