Show only 'active' repos for search and owner/member
Currently we don't have a good way of checking if hooks are on or of for
a repo, because people can add them manually. That's why we use simple
check for builds - if builds exist, it means that repo is active.

Recently (3ebe410) I made a change that removed scope for that
(timeline) from /repos endpoint, unless it's the default request,
without any other modifiers. It turns out that it's not a good idea,
because it may show irrelevant repos when doing search or viewing "My
Repositories" tab.