Add a repository_slug to jobs payload
When we fetch jobs for queues, we need also repository data to show
links. The easiest way to not do this is to attach repository slug to
job's payload. Ideally, we should not need such one purpose fields, but
for it's the best optimization that we can do without too much effort at
the moment.