Take JSON conversion into account when splitting log into chunks
If we split log into chunks checking the byte size, chunks may be still
too big after converting them to json. Splitting JSON in a way we keep
it valid is not that easy, so for now I use much simpler, but not so
nice method: try to append parts of log to a chunk and create a new
chunk as soon as adding the part would make chunk too big.