Remove :rescues and :new_relic from Task#run
We don't use newrelic for tasks anymore since extracting it.

Exceptions should bubble up to Sidekiq so tasks can be retried. We might
re-add this and make it re-raise the exception instead so that we'd get
exceptions reported. Also, we'd probably want to add task-specific
filtering and handling for exceptions (such as invalid email addresses
or inresponsive webhooks) in the tasks.