Change method of authenticating the request from github
After introducing DB encryption, I've changed authenticate_by to search
for encrypted string. Although it works with some of the AES modes, it
will certainly not work with CBC. Because of that we need to get the
user from database by given login and check the token by fetching them
from the database.