Warn when log part needs to be split.
In order to make handling log parts on the client easier we don't want
to handle numbers like "4.1". travis-worker will be changed to send
payload lower than 9kB to fit nicely into pusher message size limit and
then we should not generally need chunkifier in travis-core. To ensure
that this is true I'm leaving chunkifier in place and I added warning
triggered when it needs more than one chunk.