Give 100B of extra space to chunks
We still sometimes get chunks bigger than 9kB from workers. This is hard
to debug at the moment and they're never more than 10-20B over the
limit. This should make them fit and prevent splitting it for pusher and
there will be still a lot of space for other fields in JSON.