Don't mix global env vars with matrix env vars
After introduction of global environment varialbles they were mixed into
each matrix :env values. Such way of dealing with them causes a range of
problems, mostly related to the fact that global env vars are not
actually part of the matrix, but are included as matrix elements.

This commit changes this, to send global env vars as
config[:global_env] and not mix them with matrix.