When sending global_env in config is disabled, don't send it!
I was using global_env key in config hash in order to fix a bug with
allow_failures. Later on I also added explicit support for sending
global_env to workers. The global_env key used for fixing the bug,
should not go to workers, because it has a slightly different format
than workers can handle.

That's why it needs to be saved in config as something else, like
:_global_env instead of :global_env