Don't check for last_build_started_at when getting for repos with builds
When we get repos for recent repos sidebar, we want only repos with
either started or finished builds. It looks differently on 'my
repositories' tab or when you search for repos. We should not display
repos without any builds there, because that most of the time means that
they're not configured, but we should not ommit a repo only because its
last build is queued.