Specify table when using select with * and other tables joined in
When 2 tables are joined in a query, rails will not automatically choose
the proper columns.

This commit does not have tests, because it fixes broken tests from
previous commits, which appeared only when running the entire test
suite, I want to leave it as separate commit, so it hopefully gives any
value to someone - it's somewhat obvious when you know what's going on,
but it took me quite some time to figure out what's going on.