Remove uniquness constraint on owner and name on repositories
We sometimes rely on the fact that repos have unique name and owner_name
(for example when user opens a repo page, like: /rails/rails), but there
can be a situation when repo is renamed on github and the new one is
created with the previous name. In such situation we should allow to create
the new repository - the name will be duplicated, but this will last
only till we sync the DB with github.