Add cancelable to build and job payloads
In order to not move too much business logic, I added cancelabke?
attribute to all build's and job's payloads. This will increase the
amount of data we send, but on the other hand it will greatly reduce a
cost of changing cancel logic (which I'm sure will be changed soon, as
it handles only the simplest scenario at the moment)