Revert "Add cancelable to build and job payloads"
This reverts commit 03ddc1ee0bfc37d7a900376ae7c48eeb97dbbaef.

The idea of encapsulating logic in the API is nice, but after all our
live updates model is not ideal for such task. The problem is that we
won't send updates to build if cancelable changes, because it may happen
without changing state of the build.

For example: a build has 2 jobs. One job is in created state and the
other is in passed. Currently, a build is cancelable when any of its
jobs is cancelable, so in such situation the build is cancelable. When
state of the created job changes to queued, its no longer cancelable and
thus build should also be not cancelable, but because there was no
change of the build's state, pusher update won't be sent.

Ideally, in the future, we should make pusher updates smarter, but for
now it will be hard to pursue having only one place for such logic.