Properly sort repositories on the lists
`timeline` scope returns repositories sorted by `last_build_started_at`,
but it also does not return repositories, which are queued. I don't want
to change that for now, but I want repositories with `owner_name` and
`member`  queries to be sorted by the start date of last build (with
queued repositories in the front).