Raise error when we don't have repository data in the payload
We got some payload with missing repository data recently and we're not
sure yet what went wrong. @github suggested to log delivery guid and
event type for such request, so we can get more information about it.

This commit raises an error with delivery guid and event type
when repository data is missing