New repo settings option: build PRs on synchronize
When new commit is pushed to a target branch, github sends us
synchronize event. We haven't tested this commit, unless head was
changed, but it is a valid use case - when master changes, PR may no
longer pass, so it's good to retest it.

This commit allows to enable such a behavior by changing a
`builds.build_pull_requests_on_synchronize?`