Expand build matrix only on relevant factors
Build matrices should be constructed in such a way that
only relevant ones are built; for example, if the language
is 'python', it does not make sense to expand on 'rvm' or 'gemfile'.

These choices are hardcoded in `Build::EXPANSION_KEYS_LANGUAGE`.
We always expand on the keys `Build::EXPANSION_KEYS_UNIVERSAL`.

Also, eradicate use of `Build.expand_matrix_config`.