Move checking for build_pushes? and build_pull_requests? to Approval
Rejecting a pull request or a push in receive service results in not
creating a request object. This is not good when pushes or pull requests
are disabled in settings, because we have no way to tell what really
happend. For example: someone disables pushes in repo settings, another
person pushes, request nor build is not created. We can assume that it's
because of the settings, but currently we would need to check logs in
order to be sure.

This commit moves this check to Approval, which will also make it easier
to communicate a reason to the user.