Log detailed build matrix information when matrix is in invalid state
The original motivation for accounting for 'created', 'queued' and
'started' states was to eradicate 'Invalid job state' errors we were
seeing in the wild, but upon further reflection, we believe this
should not happen.

Unfortunately, we are not sure about the real reason for the invalid
state, so log more information about the matrix instead, in order to
gain more insight.