Add Repository::Settings::Collection and Repository::Settings::Model
This is a first stab at structuring settings API. Model and Collection
will be used in Settings to save and load data for given type (like
ssh_keys or notifications). This will allow us to keep the clean API on
top of settings without tieing us closely with underlying
representation.