Eradicate Job::Test#waiting_for_result?
Remove the confusingly named Job::Test#waiting_for_result?.
The method was returning the opposite of what the method name implied.
Removing this from Build::Matrix#matrix_finished? on the master
trips one spec that shouldn't.

https://github.com/travis-ci/travis-core/blob/dbb57d05dc1132512e93f1e789109b8bad9c56a0/spec/travis/model/build/matrix_spec.rb#L49-L57

This leads me to believe that filtering that the previous commit
introduced might already be covered by the existing spec.