sync: reset the GitHub OAuth token on sync
This causes the GitHub OAuth token to only be valid for about 24 hours,
and should there be a leak at one point, the token will only be valid
for a limited amount of time.

This could break other code that accesses the API at the same time as
the sync is being run, since the old OAuth token is immediately revoked.