Try to use update_column instead of update_attributes
This is a try to fix #1856 (API shows old log if the log is archived to
s3). The TODOs claim that update_attributes does not work for
aggregated_at which may be also true for archived_at. Tests seem to work
with both versions, but I will try this change anyway.