travis-yml-stats: pass some more information to Keen IO
- event_type so we can track pull requests vs regular pushes
- matrix size so we can see how many jobs repositories test against
- owner_id/owner_type/owner so we can do a unique count on owners,
  similar to how we can do that for repositories

The reason for owner_id, owner_type and owner all being sent is
partially due to our database schema and partially due to the querying
API of Keen IO. First of all, we need both owner_id and owner_type since
the ID sequences of users and organizations are different, so the id
alone is not a guarantee of uniqueness. Second, we cant group by more
than one parameter, so we need the array/tuple form for grouping, but
the "normalized" version for querying directly on the values (useful for
looking at other values for organizations versus users).