Simplify Requests::Service::Receive
In anchient times users would set up their service hooks on GitHub
manually. We then created the repository and owner instances on
our side based on the incoming request data.

This is now outdated. We require people to set the service hook on
our side, and the repository and owner records will be created at
user sync time.