change to lock -> transaction -> work flow
All of the transaction types supplied by prostgres have an issue where as soon as you do one query inside of the transaction, be it setting the transaction type, or checking for the advisory lock, then all queries made within the transaction will only see results based on data from before the transaction was started.

In the case of incrementing build numbers, if two transactions start at the same time, then the advisory locks kick in, both will see the old next build number for the repository, and thus both will have the same number once the request is processed.

This change will help the situation though, as the transaction won’t start until the advisory lock can be obtained.