Coerce `Resporitoy#regenerate_key!` to really regenerate the key
Unfortunately, we ended up with a few thousand repositories
without SSL key, because we changed the way Repository models
are instantiated.
(Basically, we bypassed the ActiveRecord hooks to generate them.)

There is no simple way to force key regeneration at the moment,
so we brute force the key generation by going into the console,
select the suspect Repositories and call `#regenerate_key!`
on them.