job-queue: allow random routing by percentage
In some cases we want to send a subset of jobs to a new queue, usually
for testing out a new infrastructure of some sort. This allows
specifying a percentag (in the range [0, 100)), and then the jobs will
be distributed with weighted randomness. For example, consider this
configuration:

- queue: builds.new-a
  language: a
  percentage: 25
- queue: builds.old-a
  language: a

If you have 100 jobs with `language: a`, roughly 25 of them will go to
builds.new-a, and the rest will go to builds.old-a.

The way the matching works is by stepping through each config from top
to bottom, and applying the first one that matches. With this addition,
the `percentage` key will cause it to be possible for the matching to
continue to the next config even if the job matched the rest of the
config.