Fix StatesCache::EventHandler
In a5ff36f I changed a way we check if states cache needs to be updated.
The problem was that we check for the newest build by id in the DB, but
StatesCache class used finished_at. Addons::StatesCache::EventHandler
wasn't updated along with StatesCache and it still only sent finished_at
instead of build's id. This commit adds a test for
Addons::StatesCache::EventHandler and fixes the issue.