improve error handling for event subscriptions looking at missing addons
With a certain event handler present in the config, but missing in the
code, our previous code would catch and log the fact that the constant
is missing. However, the method now also would return `true` (the return
value of `logger.error`), causing the caller to try to find a constant
on the value `true`, which of course would raise a new exception. This
is confusing, and not the behaviour one would expect: broken event
listener configuration should not break the application in general.

This change fixes that, changes the return value for `nil` in the
offending case, and deals with this value appropriately.