Remove intermediate `#bucket` call
Buckets are easy to work with in S3, but for GCS, this is not the
case.
For example, it is not quite easy (if not impossible) to
get a list of objects belonging to the bucket.
Instead, one has to do:

    storage.list_objects(bucket_name, prefix: prefix).items

(This part of GCS logic is not yet implemented.)